# This is a combination of 2 commits. # The first commit's message is:
Updated RDS transmitter: flags, PI and date/time

Merging baseband audio tone generators

Merging DTMF baseband with "tones" baseband

Added stealth transmit mode

App flash section bumped to 512k
RX and TX LEDs are now used
Play dead should work again, added login option
Morse frame gen. for letters and fox hunt codes
Merged EPAR with Xylos
Made EPAR use encoders for frame gen.
Moved OOK encoders data in encoders.hpp
Simplified about screen, ui_about_demo.* files are still there

BHT city DB, keywords removed

BHT cities DB, keywords removed

Update README.md

RDS radiotext and time group generators

# This is the 2nd commit message:

Update README.md

+++ firmware/application/ui_textentry.cpp
@@ -24,7 +24,7 @@
 
 namespace ui {
 
-bool textentry(NavigationView& nav, char * str, uint16_t max_length) {
+bool textentry(NavigationView& nav, char * str, size_t max_length) {
 	if (portapack::persistent_memory::ui_config_textentry() == 0) {
 		auto an_view = nav.push<AlphanumView>(str, max_length);
 		an_view->on_changed = [str, max_length](char * value) {
